Move the Path helper into a separate file
