+++ Plugins/SwiftLintPlugin/Path+Helpers.swift
@@ -1,0 +1,18 @@
+import Foundation
+import PackagePlugin
+
+extension Path {
+    /// Scans the receiver, then all of it's parents looking for a configuration file with the name ".swiftlint.yml".
+    /// - Returns: Path to the configuration file, or nil if one cannot be found.
+    func firstConfigurationFileInParentDirectories() -> Path? {
+        // TODO: Consider linking to the framework to get the default configuration file name
+        let defaultConfigurationFileName = ".swiftlint.yml"
+        let proposedDirectory = sequence(first: self, next: { $0.removingLastComponent() })
+            .first { path in
+                let potentialConfigurationFile = path.appending(subpath: defaultConfigurationFileName)
+                return FileManager.default.isReadableFile(atPath: potentialConfigurationFile.string)
+            }
+
+        return proposedDirectory?.appending(subpath: defaultConfigurationFileName)
+    }
+}
