+++ Plugins/SwiftLintPlugin/SwiftLintPlugin.swift
@@ -90,19 +90,3 @@ extension SwiftLintPlugin: XcodeBuildToolPlugin {
     }
 }
 #endif
-
-private extension Path {
-    /// Scans the receiver, then all of it's parents looking for a configuration file with the name ".swiftlint.yml".
-    /// - Returns: Path to the configuration file, or nil if one cannot be found.
-    func firstConfigurationFileInParentDirectories() -> Path? {
-        // TODO: Consider linking to the framework to get the default configuration file name
-        let defaultConfigurationFileName = ".swiftlint.yml"
-        let proposedDirectory = sequence(first: self, next: { $0.removingLastComponent() })
-            .first { path in
-                let potentialConfigurationFile = path.appending(subpath: defaultConfigurationFileName)
-                return FileManager.default.isReadableFile(atPath: potentialConfigurationFile.string)
-            }
-
-        return proposedDirectory?.appending(subpath: defaultConfigurationFileName)
-    }
-}
